# src/lambda/swagger.yaml
openapi: 3.0.0
info:
  title: Basilwizi Trust Home Page
  description: Basilwizi Trust, people of the Great river. They are found along the Kasambabezi River
  version: 1.0.0
servers:
  - url: https://localhost:7150
    description: Production Server
paths:
...
  /subscribe/saveemail:
    post:
      summary: Save the email of subscriber
      description: Save emails of subscribers
      operationId: saveEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "basilwizi@basilwizi.org"
              required:
                - email
      responses:
        "200":
          description: The email has been successfully subscribed to the Basilwizi and will receive regular updates on projects and latest stories
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Subscriber added to the database successfully"
...
  /likes:
    get:
      summary: Get a list of all like
      description: No restriction.
      operationId: getAllLikes
      responses:
        "200":
          description: An array of all likes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "5eb12e197e06a76ccdefc121"
                    liked:
                      type: boolean
                      example: true
                    liked_by:
                      type: string
                      example: "Basil Lwizi"
                    like_to:
                      type: string
                      example: "5eb12e197e06a76ccdefc121"
                    created:
                      type: string
                      example: "2020-05-05T09:12:57.848Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /likes/{id}/like:
    post:
      summary: Like a post or article
      description: Like a post by another writer
      operationId: likePost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                liked:
                  type: boolean
                  example: true
                liked_by:
                  type: string
                  example: "Basil Lwizi"
              required:
                - liked
                - liked_by
      responses:
        "200":
          description: Thank you for liking a post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  liked:
                    type: boolean
                    example: true
                  liked_by:
                    type: string
                    example: "Basil lwizi"
                  like_to:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
        "400":
          description: There was an error saving to DB
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "An error occured while saving to DB"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /likes/{id}:
    parameters:
      - in: path
        name: id
        description: Like id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a like's details
      description: Likes of posts have details
      operationId: getLikeById
      responses:
        "200":
          description: Details of the specified like
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  liked:
                    type: boolean
                    example: true
                  liked_by:
                    type: string
                    example: "Basil Lwizi"
                  like_to:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /projects/saveproject:
    post:
      summary: Create a stubnning story for the people of basilwizi to read
      description: Edit your article for Grammar and spelling before you publigh it
      operationId: saveProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_author:
                  type: string
                  example: "Basilwizi Bamulonga"
                project_title:
                  type: string
                  example: "The Great River is now producing abundant Fish for the People"
                project_description:
                  type: string
                  example: "Abuntantly available fish ends drought"
              required:
                - project_author
                - project_title
                - project_description
      responses:
        "200":
          description: Article will be sent to readers now
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The article was saved and will be available to readers"
  /projects:
    get:
      summary: Get a list of projects articles in the DB
      description: All readers have the right to these articles
      operationId: getAllProjects
      responses:
        "200":
          description: These are the articles you have
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "5eb12e197e06a76ccdefc121"
                    project_author:
                      type: string
                      example: "Basilwizi Bamulonga"
                    project_title:
                      type: string
                      example: "The Basilwizi we know"
                    project_description:
                      type: string
                      example: "The basilwizi we know is here!"
                    created:
                      type: string
                      example: "2020-05-05T09:12:57.848Z"
                    updated:
                      type: string
                      example: "2020-05-08T03:11:21.553Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /projects/{id}:
    parameters:
      - in: path
        name: id
        description: ProjectArticle id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    put:
      summary: Update an project
      description: Admin users can update projects articles and owners of articles can update their own articles.
      operationId: updateProjectArticle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_author:
                  type: string
                  example: "Basilwizi Bamulonga"
                project_title:
                  type: string
                  example: "The Great River is now producing abundant Fish for the People"
                project_description:
                  type: string
                  example: "Abuntantly available fish ends drought"
      responses:
        "200":
          description: Article updated successfully. The details of the updated account are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  project_author:
                    type: string
                    example: "Basilwizi Bamulonga"
                  project_title:
                    type: string
                    example: "The Great River is now producing abundant Fish for the People"
                  project_description:
                    type: string
                    example: "Abuntantly available fish ends drought"
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  updated:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"